lib/ukfallocbuddy: Add buddy-based frame allocator
ukfallocbuddy is a multi-zone buddy-based frame allocator. For
managing free space, the allocator uses per-level free lists and a
hierarchical non-inclusive bitmap (i.e., bits are only set on one level
for a certain range). The hierarchy allows having to set only a single
bit also for large allocations such as 1 GiB. In contrast to a
regular buddy allocator, ukfallocbuddy allows allocations that are
not a power-of-two without wasting the remaining space as internal
fragmentation (e.g., a 6-frames allocation retrieves an 8-frames buddy
area and gives back the last two frames to the allocator). Smallest
allocation size is 1 frame. The largest (continuous) allocation size
is configurable.

The allocator is optimized for fast allocation from any address (O(1))
but also supports direct allocation of certain physical address ranges
and allocation from a limited physical address range. The latter have
a complexity of O(n) with n being the number of entries in the free
lists. As n is rather small early in the boot phase (i.e., low external
fragmentation), direct and restricted allocations should be done
in this phase for best performance. Keeping free lists instead of
trees is justified with optimizing for the average case, which is a
power-of-two any-address allocation that can be served by lists with
minimal overhead.

The allocator reserves around 64KiB for metadata for each GiB of
assigned physical memory (mostly the bitmap). However, if the
paging implementation does not provide direct-mapped physical
memory (i.e., `CONFIG_PAGING_HAVE_DIRECTMAP`), per-frame
metadata must be stored explicitly which adds 24 bytes per frame,
which is an additional ~6 MiB for each GiB of assigned physical
memory. See the code for a comment on how to reduce this to ~2 MiB
if needed.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>