plat/common/x86: Adapt exception handling for crashdump
Introduce UK_EVENT_UNHANDLED_EXCEPTION, invoked by the fallback
trap handler, do_unhandled_trap(). This enables registering a
a custom crashdump handler. The default implementation is on
libukdebug.

Remove explicit page fault handler, as with the addition of the
unhandled exception event it no longer provides additional value
compared to the handler generated by DECLARE_TRAP_EC(). Instead,
pass the value of CR with the potentially faulting address to the
trap context.

Amend the trap context with a human-readable description of the
exception and a system error. The system error should be populated
by handlers.

Remove old register dump.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>