lib/ukdebug: Add crash screen
The crash screen is registered as a handler of the latest priority
for the UNHANDLED_EXCEPTION event, to provide a crashdump before
executing a predefined action. The default action is shutdown, and
reboot is supported conditionally to CONFIG_HAVE_REBOOT.

The crash screen additionally provides a CRASH event, with high-level
information on the cause behind the system failure.

Checkpatch-Ignore: LONG_LINE
Checkpatch-Ignore: QUOTED_WHITESPACE_BEFORE_NEWLINE
Signed-off-by: Michalis Pappas <michalis@unikraft.io>
Co-authored-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Co-authored-by: Marco Schlumpp <marco@unikraft.io>