drivers/virtio/mmio: Support devices in command line parameters
This commit adds support for defining virtio-mmio devices in the
command line parameters.

The probing function is invoked by ukbus at a priority lower than
the the platform bus, to ensure that the virtio devices have been
probed, to safely register with the parent device.

Add a config parameter to control compatibilty with Linux.
Specifically, when VIRTIO_MMIO_LINUX_COMPAT_CMDLINE is selected
the MMIO driver is configured to accept devices in the Linux format,
ie:

  virtio_mmio.device = <size>@<base>:<irq>[:<id>]

If this option is not enabled, use the defult libukparam
format for this library, ie:

  libvirtio_mmio.device = <size>@<base>:<irq>[:<id>]

This is useful for VMMs like Firecracker, that automatically inject
device descriptors the command line.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>