plat/common: Rework initialzation of paged memory
This commit reworks the initialization of paged memory to provide a
more flexible implementation that is capable of handling regions
beyond the limits of those defined in the boot pagetables. The
motivation for this change is to allow mapping device regions that are
unknown at compile-time, such as Unprotected IPA Alias regions of Arm
CCA Realms, the address of which depends on the executing platform.

Under the new scheme the concept of bootinfo is simplified to only
contain valid memory regions. Upon paged memory initialization these
regions are either mapped with the protections defined in their mrd,
or are already mapped, in which case only their protections are
updated. An exception to this is mrds of the UKPLAT_MEMRT_DEVICE type,
that are ignored to allow MMIO regions mapped during early device
initialization to remain mapped.

Any memory not defined in bootinfo is unmapped. The theoretical bounds
used as a reference are platform-specific and are defined in the newly
introduced platform.h.

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Michalis Pappas <michalis@unikraft.io>