lib/syscall-shim: Do not dereference optional clone() parameters
The parent_tid parameter of clone() is used to instruct the kernel
where to store the child TID in parent's memory. Similarly, the
child_tid is used to instruct the kernel where to store the child
TID in the child's memory. Both parameters are optional.

Do not interpret these options as PT_REF, as it causes a memory fault
when the caller passes NULL. Instead print the pointer value as PT_VADDR.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>